requirements.txt
aiohttp==3.9.1
websockets==12.0
python-binance==1.0.19
xgboost==2.0.3
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
structlog==23.2.0
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
simple-websocket==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
